+++ ContentPatcher/docs/release-notes.md
@@ -11,6 +11,7 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 -->
 ## Upcoming release
 * Fixed various issues caused by pre-1.6 content packs replacing data assets which changed format in 1.6.
+* Fixed runtime migrations for pre-1.6 content packs letting them remove 1.6 content.
 
 ## 2.0.5
 Released 07 April 2024 for SMAPI 4.0.0 or later.